Rename: getGeolocation -> geolocation 파일명 변경
